+++ sdks/outpost-typescript/src/funcs/destinationsList.ts
@@ -113,7 +113,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc("/tenants/{tenant_id}/destinations")(pathParams);
 
   const query = encodeFormQuery({