+++ .speakeasy/logs/changes/changes.md
@@ -17,3 +17,43 @@
     - `ResponseData` **Changed** (Breaking ⚠️)
 * `Outpost.Metrics.GetEventMetrics()`: **Added**
 * `Outpost.Metrics.GetAttemptMetrics()`: **Added**
+
+## Typescript SDK Changes:
+* `outpost.attempts.list()`: `response.models[]` **Changed** (Breaking ⚠️)
+    - `destination` **Added**
+    - `event` **Changed** (Breaking ⚠️)
+    - `responseData` **Changed** (Breaking ⚠️)
+* `outpost.attempts.get()`: `response` **Changed** (Breaking ⚠️)
+    - `destination` **Added**
+    - `event` **Changed** (Breaking ⚠️)
+    - `responseData` **Changed** (Breaking ⚠️)
+* `outpost.destinations.listAttempts()`: `response.models[]` **Changed** (Breaking ⚠️)
+    - `destination` **Added**
+    - `event` **Changed** (Breaking ⚠️)
+    - `responseData` **Changed** (Breaking ⚠️)
+* `outpost.destinations.getAttempt()`: `response` **Changed** (Breaking ⚠️)
+    - `destination` **Added**
+    - `event` **Changed** (Breaking ⚠️)
+    - `responseData` **Changed** (Breaking ⚠️)
+* `outpost.metrics.getEventMetrics()`: **Added**
+* `outpost.metrics.getAttemptMetrics()`: **Added**
+
+## Python SDK Changes:
+* `outpost.attempts.list()`: `response.models[]` **Changed** (Breaking ⚠️)
+    - `destination` **Added**
+    - `event` **Changed** (Breaking ⚠️)
+    - `response_data` **Changed** (Breaking ⚠️)
+* `outpost.attempts.get()`: `response` **Changed** (Breaking ⚠️)
+    - `destination` **Added**
+    - `event` **Changed** (Breaking ⚠️)
+    - `response_data` **Changed** (Breaking ⚠️)
+* `outpost.destinations.list_attempts()`: `response.models[]` **Changed** (Breaking ⚠️)
+    - `destination` **Added**
+    - `event` **Changed** (Breaking ⚠️)
+    - `response_data` **Changed** (Breaking ⚠️)
+* `outpost.destinations.get_attempt()`: `response` **Changed** (Breaking ⚠️)
+    - `destination` **Added**
+    - `event` **Changed** (Breaking ⚠️)
+    - `response_data` **Changed** (Breaking ⚠️)
+* `outpost.metrics.get_event_metrics()`: **Added**
+* `outpost.metrics.get_attempt_metrics()`: **Added**